+++ src/commands/git_secret_removeperson.sh
@@ -1,14 +1,14 @@
 #!/usr/bin/env bash
 
 
-function killperson {
+function removeperson {
   OPTIND=1
 
   while getopts 'h' opt; do
     case "$opt" in
-      h) _show_manual_for 'killperson';;
+      h) _show_manual_for 'removeperson';;
 
-      *) _invalid_option_for 'killperson';;
+      *) _invalid_option_for 'removeperson';;
     esac
   done
 
@@ -22,7 +22,7 @@ function killperson {
   local emails=( "$@" )
 
   if [[ ${#emails[@]} -eq 0 ]]; then
-    _abort "at least one email is required for killperson."
+    _abort "at least one email is required for removeperson."
   fi
   # Getting the local git-secret `gpg` key directory:
   local secrets_dir_keys
@@ -42,4 +42,10 @@ function killperson {
   _message 'removed keys.'
   _message "now [$*] do not have an access to the repository."
   _message 'make sure to hide the existing secrets again.'
+}
+
+function killperson {
+  echo 'Warning: `killperson` has been renamed to `removeperson`. This alias will be removed in the future versions, please switch to call `removeperson` going forward.'
+
+  removeperson "$@"
 }